==> 1_Descriptive_Statistics.ipynb <==
# Company's product area, revenue and customer base trends
### Summary of I/O csv files
Input Files
    > 'area.csv'     - product area code & descripton
    > 'caregory.csv' - product sub-category...
Output Files
    > 'company_area_percentage.csv'         - percentage revenue for all companies by ProductArea
    > 'company_category_percentage.csv'     - ... by MajorCategory
    > 'company_sales_revenue_36.csv'        - revenue for Company 36 in 36 months in 10 areas
    > 'company_sales_hh_num_36.csv'         - cumulative customer base for ...

## Use MongoDB to Store Source Data
Start MondoDB as a Windows Service from command line:

> $ "C:\Program Files\MongoDB\Server\3.4.2\bin\mongod.exe" --config "C:\Program Files\MongoDB\Server\3.4.2\mongod.cfg" --install

> $ net start MongoDB

Import csv files to db BT4221_DB 
> $ cd C:\Program Files\MongoDB\Server\3.4.2\bin

> $ mongoimport -d BT4221_DB -c Orders --type csv --file C:\Users\User\Desktop\BT4221\Project\11\DMEFOrders3Dataset2.csv --headerline

> $ mongoimport -d BT4221_DB -c Lines --type csv --file C:\Users\User\Desktop\BT4221\Project\11\DMEFLines3Dataset2.csv --headerline

> $ mongoimport -d BT4221_DB -c HouseholdZIP --type csv --file C:\Users\User\Desktop\BT4221\Project\11\4-DMEF3YrBase.csv --headerline


# Python version used
import sys
sys.version

""" Method for retriving data from mongodb"""
import pandas as pd
from pymongo import MongoClient

def _connect_mongo(host, port, username, password, db):
    """ A util for making a connection to mongo """
    if username and password:
        mongo_uri = 'mongodb://%s:%s@%s:%s/%s' % (username, password, host, port, db)
        conn = MongoClient(mongo_uri)
    else:
        conn = MongoClient(host, port)
    return(conn[db])

def read_mongo(db, collection, query={}, host='localhost', port=27017, username=None, password=None, no_id=True):
    """ Read from Mongo and Store into DataFrame """
    # Connect to MongoDB
    db = _connect_mongo(host=host, port=port, username=username, password=password, db=db)
    cursor = db[collection].find(query)
    df =  pd.DataFrame(list(cursor))
    # Delete the _id
    if no_id:
        del df['_id']
    return(df)

# 
## Describe all companies by product areas and subcategories

lines = read_mongo("BT4221_DB", "Lines")
print(len(lines))

""" describe all companies by product area"""
lines.ProductArea = lines.ProductArea.apply(lambda s: s if s<10 else 10) # 10 product areas
df_area = lines.groupby(by=["CompanyID", "ProductArea"])['Dollars'].sum().to_frame().reset_index()
# pivot the dataframe
df_area_wide = df_area.pivot(index='CompanyID', columns='ProductArea', values='Dollars')

# change to % by revenue
import numpy as np
df_area_wide = df_area_wide.replace(np.nan, 0, regex=True)
s = df_area_wide.sum(axis=1)
df_area_perc = df_area_wide.loc[:,:].div(s, axis=0)

# change column names to actual areas
area = pd.read_csv("area.csv")
names = area.Description.head(10).tolist()
names = names[:-1]+["Others"]
df_area_perc.columns = names
print(df_area_perc.shape)
df_area_perc.head(2)

# Save results to csv file
file_name = "company_area_percentage.csv"
df_area_perc.to_csv(file_name)

print(df_area_perc.shape)
df_area_perc.sample(1)

print(df_area_perc.shape)
df_area_perc.sample(1)

""" company by product category"""
lines.MajorCategory = lines.MajorCategory.apply(lambda s: s if s<73 else 2) # change 1 and 73 to 2
lines.MajorCategory = lines.MajorCategory.apply(lambda s: s if s>1 else 2) # change 1 and 73 to 2

df_category = lines.groupby(by=["CompanyID", "MajorCategory"])['Dollars'].sum().to_frame().reset_index()
# pivot the dataframe
df_category_wide = df_category.pivot(index='CompanyID', columns='MajorCategory', values='Dollars')

# change to % by revenue
import numpy as np
df_category_wide = df_category_wide.replace(np.nan, 0, regex=True)
s = df_category_wide.sum(axis=1)
df_category_perc = df_category_wide.loc[:,:].div(s, axis=0)

# change column names to actual categorys
category = pd.read_csv("category.csv")
df_category_perc.rename(columns=dict(zip(category.Code, category.Description)), inplace=True)
print(df_category_perc.shape)
df_category_perc.sample(1)

# Save to csv files
file_name = "company_category_percentage.csv"
df_category_perc.to_csv(file_name)

#
## Identify a company 
> with relative large number of unique customers in the training period

df1 = lines[lines.OrderDate<=20070630].groupby('CompanyID').agg({"HH_ID": 'nunique', 'Dollars': 'sum'})
df1 = df1.sort_values(by="HH_ID", ascending = False)
df1.head(3)

### Thus, we select CompanyID = 36 as an example
# 
## Trends for Company 36
> Trend for Revenue in Each Product Area, by month

> Trend for cumulative customer base in each ProductArea, by month

company = 36
df = lines[lines.CompanyID==company]
df["month"] = df.OrderDate//100
print("number of lines:", len(df))
print("period: ",min(df.month),max(df.month))
df.sample(1)

"""group by month, 10 product areas, revenue as sum dollars"""
df2 = df.groupby(by=["ProductArea", "month"])['Dollars'].sum().to_frame().reset_index()
df2.sample(1)
# Save to csv
file_name = "company_sales_revenue_"+str(company)+".csv"
df2.to_csv(file_name, index=False)

"""group by month, 10 product areas, accumulative customer base = count hh_id"""
# For each month-ProductArea, count the number of unique customers in accumulative customer base
def num_hh_count(s):
    df_area = df.groupby('ProductArea').get_group(s.ProductArea)    
    df_month = df_area[df_area.month<=s.month]    
    hh_num = len(set(df_month.HH_ID)) # or df_month.HH_ID.nunique()
    return(hh_num)
df2["num_hh"] = df2.apply(num_hh_count, axis=1) 
df3 = df2[["ProductArea", "month", "num_hh"]]
df3.sample(1)
# Save to csv
file_name = "company_sales_hh_num_"+str(company)+".csv"
df3.to_csv(file_name, index=False)df3.sample(1)
