==> 2_CF_DataPrep.ipynb <==
## Data required by recomender
> in the format of: user item rating; tab deliminated, withut header

> train/test: MinMaxScalling or not; Number of purchases or Total spending in Purchases; log trascormation or not;

16 output files with naming convension: 
     >   [scale_,][train_, test_][num, dol][_log,].txt
     

""" Retriving data from mongodb"""
import pandas as pd
from pymongo import MongoClient

def _connect_mongo(host, port, username, password, db):
    """ A util for making a connection to mongo """
    if username and password:
        mongo_uri = 'mongodb://%s:%s@%s:%s/%s' % (username, password, host, port, db)
        conn = MongoClient(mongo_uri)
    else:
        conn = MongoClient(host, port)
    return(conn[db])

def read_mongo(db, collection, query={}, host='localhost', port=27017, username=None, password=None, no_id=True):
    """ Read from Mongo and Store into DataFrame """
    # Connect to MongoDB
    db = _connect_mongo(host=host, port=port, username=username, password=password, db=db)
    cursor = db[collection].find(query)
    df =  pd.DataFrame(list(cursor))
    # Delete the _id
    if no_id:
        del df['_id']
    return(df)

orders = read_mongo("BT4221_DB", "Orders")
print(len(orders)) 
orders.sample(2)

"""training data, first 30 months, HH_ID-CompanyID level data"""
train = orders[orders.OrderDate<=20070630].groupby(by=["HH_ID", "CompanyID"]).agg({"OrderNum": 'nunique', 'Dollars': 'sum'})
train.reset_index(inplace=True)
print(len(train))
train.sample(1)

"""rating as log() """
import numpy as np
train["logDollar"] = np.log(train.Dollars)
train["logOrder"] = np.log(train.OrderNum)
train.sample(1)

"""MinMax Scalling"""
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler(feature_range=(1, 5))
cols = ["OrderNum","Dollars","logOrder","logDollar"]
scale_train = train.copy()
scale_train[cols] = scaler.fit_transform(scale_train[cols])
scale_train.sample(1)

""" save as format required by recommender library"""
for col in ["OrderNum","Dollars","logOrder","logDollar"]:
    cols = ["CompanyID","HH_ID", col]
    measure = "dol" if "Dollar" in col else "num"
    transformation = "_log" if "log" in col else ""
    file_name = "train_" + measure + transformation + ".txt"
    print(file_name)
    train[cols].to_csv(file_name, index=False, header=False,sep="\t", quoting=csv.QUOTE_NONE )
    scale_train[cols].to_csv("scale_"+file_name, index=False, header=False,sep="\t", quoting=csv.QUOTE_NONE )

""" Customers contributing to company 36's revenue in testing period """
company = 36
test = orders[(orders.OrderDate>20070630) & (orders.CompanyID==company)].groupby(by=["HH_ID", "CompanyID"]).agg({"OrderNum": 'nunique', 'Dollars': 'sum'})
test.reset_index(inplace=True)
print(len(test))
test.sample(1)

""" Recomended Customers """
recommended_hh = pd.DataFrame({'HH_ID': list(set(train.HH_ID)-set(train[train.CompanyID==company].HH_ID))})
test_output  = pd.merge(recommended_hh,test,how='left', on = ["HH_ID"])
test_output["HH_ID"] = test_output["HH_ID"].astype(int).astype('str')
test_output["CompanyID"] = str(company)
test_output = test_output.fillna(0).sort_values(by="Dollars",ascending=False).reset_index(drop=True)
print(len(test_output))
test_output.head(3)

"""rating as log(measure+1) """
import numpy as np
test_output["logDollar"] = np.log(test_output.Dollars+1) # To prevent -inf from appearing
test_output["logOrder"] = np.log(test_output.OrderNum+1)
test_output.sample(1)

"""MinMax Scalling """
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler(feature_range=(1, 5))
cols = ["OrderNum","Dollars","logOrder","logDollar"]
scale_test_output = test_output.copy()
scale_test_output[cols] = scaler.fit_transform(scale_test_output[cols])
scale_test_output.sample(1)

""" save to format required by recommender library"""
for col in ["OrderNum","Dollars","logOrder","logDollar"]:
    cols = ["CompanyID","HH_ID", col]
    measure = "dol" if "Dollar" in col else "num"
    transformation = "_log" if "log" in col else ""
    file_name = "test_" + measure + transformation  + ".txt"
    print(file_name)
    test_output[cols].to_csv(file_name, index=False, header=False,sep="\t", quoting=csv.QUOTE_NONE )
    scale_test_output[cols].to_csv("scale_"+file_name, index=False, header=False,sep="\t", quoting=csv.QUOTE_NONE )

""" save as format required by recommender library"""
for col in ["OrderNum","Dollars","logOrder","logDollar"]:
    cols = ["CompanyID","HH_ID", col]
    measure = "dol" if "Dollar" in col else "num"
    transformation = "_log" if "log" in col else ""
    file_name = "train_" + measure + transformation + ".txt"
    print(file_name)
    train[cols].to_csv(file_name, index=False, header=False,sep="\t", quoting=csv.QUOTE_NONE )
    scale_train[cols].to_csv("scale_"+file_name, index=False, header=False,sep="\t", quoting=csv.QUOTE_NONE )
