==> 6_Demo.ipynb <==
import plotly
import plotly.plotly as py
import pandas as pd
import numpy as np
import plotly.graph_objs as go
import os

**Descriptive Analysis**

# Plot Customers Household Number

df = pd.read_csv('company_sales_hh_cust_36.csv')

trace1 = go.Scatter(
    x=df['month'],
    y=df['Electronics'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(184,247,212)'),
    fill ='tonexty',
    name = 'Electronics'
)
trace2 = go.Scatter(
    x=df['month'],
    y=df['Entertainment'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(111,231,219)'),
    fill = 'tonexty',
    name = 'Entertainment'
)
trace3 = go.Scatter(
    x=df['month'],
    y=df['Fashion'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(167,166,238)'),
    fill ='tonexty',
    name = 'Fashion'
)
trace4 = go.Scatter(
    x=df['month'],
    y=df['Food & Entertaining'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(71,190,241)'),
    fill = 'tonexty',
    name = 'Food & Entertaining'
)
trace5 = go.Scatter(
    x=df['month'],
    y=df['Health & Beauty'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(124,247,212)'),
    fill ='tonexty',
    name = 'Health & Beauty'
)
trace6 = go.Scatter(
    x=df['month'],
    y=df['Home & Garden'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(51,231,219)'),
    fill = 'tonexty',
    name = 'Home & Garden'
)
trace7 = go.Scatter(
    x=df['month'],
    y=df['Kids & Baby'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(67,166,238)'),
    fill = 'tonexty',
    name = 'Kids & Baby'
)
trace8 = go.Scatter(
    x=df['month'],
    y=df['Teens'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(71,90,241)'),
    fill = 'tonexty',
    name = 'Teens'
)
trace9 = go.Scatter(
    x=df['month'],
    y=df['Travel & Sports'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(127,166,118)'),
    fill = 'tonexty',
    name = 'Travel & Sports'
)
trace10 = go.Scatter(
    x=df['month'],
    y=df['Others'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(131,90,121)'),
    fill = 'tonexty',
    name = 'Others'
)

data = [trace1, trace2, trace3, trace4, trace5, trace6, trace7, trace8, trace9, trace10]
layout = go.Layout(showlegend=True, xaxis=dict(type='category'), yaxis=dict(type='linear'))
fig = go.Figure(data=data, layout=layout)
py.iplot(fig,filename='Customer Number by Area')

# Plot Household Monthly Spending

df1 = pd.read_csv('company_sales_revenue_36.csv')

trace1 = go.Scatter(
    x=df1['month'],
    y=df1['1'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(184,247,212)'),
    fill ='tonexty',
    name = 'Electronics'
)
trace2 = go.Scatter(
    x=df1['month'],
    y=df1['1']+df1['2'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(111,231,219)'),
    fill = 'tonexty',
    name = 'Entertainment'
)
trace3 = go.Scatter(
    x=df1['month'],
    y=df1['1']+df1['2']+df1['3'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(127,166,238)'),
    fill = 'tonexty',
    name = 'Fashion'
)
trace4 = go.Scatter(
    x=df1['month'],
    y=df1['1']+df1['2']+df1['3']+df1['4'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(131,90,241)'),
    fill = 'tonexty',
    name = 'Food & Entertaining'
)
trace5 = go.Scatter(
    x=df1['month'],
    y=df1['1']+df1['2']+df1['3']+df1['4']+df1['5'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(124,247,112)'),
    fill ='tonexty',
    name = 'Health & Beauty'
)
trace6 = go.Scatter(
    x=df1['month'],
    y=df1['1']+df1['2']+df1['3']+df1['4']+df1['5']+df1['6'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(51,231,119)'),
    fill = 'tonexty',
    name = 'Home & Garden'
)
trace7 = go.Scatter(
    x=df1['month'],
    y=df1['1']+df1['2']+df1['3']+df1['4']+df1['5']+df1['6']+df1['7'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(67,166,138)'),
    fill = 'tonexty',
    name = 'Kids & Baby'
)
trace8 = go.Scatter(
    x=df1['month'],
    y=df1['1']+df1['2']+df1['3']+df1['4']+df1['5']+df1['6']+df1['7']+df1['8'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(71,90,141)'),
    fill = 'tonexty',
    name = 'Teens'
)
trace9 = go.Scatter(
    x=df1['month'],
    y=df1['1']+df1['2']+df1['3']+df1['4']+df1['5']+df1['6']+df1['7']+df1['8']+df1['9'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(227,166,118)'),
    fill = 'tonexty',
    name = 'Travel & Sports'
)
trace10 = go.Scatter(
    x=df1['month'],
    y=df1['1']+df1['2']+df1['3']+df1['4']+df1['5']+df1['6']+df1['7']+df1['8']+df1['9']+df1['10'],
    mode = 'lines',
    line = dict(width=0.5, color='rgb(231,90,121)'),
    fill = 'tonexty',
    name = 'Others'
)

data = [trace1, trace2, trace3, trace4, trace5, trace6, trace7, trace8, trace9, trace10]
layout = go.Layout(showlegend=True, xaxis=dict(type='category'), yaxis=dict(type='linear'))
fig = go.Figure(data=data, layout=layout)
py.iplot(fig,filename='Revenue by Area')

#Revenue Sources by Product Areas
area = pd.read_csv('company_area_percentage.csv')
category = pd.read_csv('company_category_percentage.csv')

table = ff.create_table(area.loc[area['CompanyID']==36])
table.layout.width = 2500
py.iplot(table, filename='Product Areas')

#Revenue Sources by Major Categories
table1 = ff.create_table(category.loc[category['CompanyID']==36])
table1.layout.width = 20000
py.iplot(table1, filename='Major Categories')

# Plot Household Spending by Area - Top 100 Households in each area

pa1 = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/hhhhhh/3.popularity of hh by spending/hh_spending_area_1_Electronics.csv")
pa2 = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/hhhhhh/3.popularity of hh by spending/hh_spending_area_2_Entertainment.csv")
pa3 = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/hhhhhh/3.popularity of hh by spending/hh_spending_area_3_Fashion.csv")
pa4 = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/hhhhhh/3.popularity of hh by spending/hh_spending_area_4_FoodEntertaining.csv")
pa5 = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/hhhhhh/3.popularity of hh by spending/hh_spending_area_5_HealthBeauty.csv")
pa6 = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/hhhhhh/3.popularity of hh by spending/hh_spending_area_6_HomeGarden.csv")
pa7 = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/hhhhhh/3.popularity of hh by spending/hh_spending_area_7_KidsBaby.csv")
pa8 = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/hhhhhh/3.popularity of hh by spending/hh_spending_area_8_Teens.csv")
pa9 = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/hhhhhh/3.popularity of hh by spending/hh_spending_area_9_TravelSports.csv")
pa10 = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/hhhhhh/3.popularity of hh by spending/hh_spending_area_10_Miscellaneous.csv")


from plotly.graph_objs import *
X = list(range(100))
electronics = pa1.iloc[:99]
entertainment = pa2.iloc[:99]
fashion = pa3.iloc[:99]
foodentertaining = pa4.iloc[:99]
healthbeauty = pa5.iloc[:99]
homegarden = pa6.iloc[:99]
kidsbaby = pa7.iloc[:99]
teens = pa8.iloc[:99]
travelsports = pa9.iloc[:99]
others = pa10.iloc[:99]

trace1 = Scatter(
    x=X, y=electronics['Dollars'],
    line=Line(
        color='F7B8B8',
        width=3
    ),
    opacity = 0.66,
    name='Electronics',
    text = electronics['HH_ID']
)

trace2 = Scatter(
    x=X, y=entertainment['Dollars'],
    line=Line(
        color='E7D06F)',
        width=3
    ),
    opacity = 0.66,
    name='Entertainment',
    text = entertainment['HH_ID']
)

trace3 = Scatter(
    x=X, y=fashion['Dollars'],
    line=Line(
        color='B1EAEF',
        width=3
    ),
    opacity = 0.66,
    name='Fashion',
    text = fashion['HH_ID']
)

trace4 = Scatter(
    x=X, y=foodentertaining['Dollars'],
    line=Line(
        color='C3BADA',
        width=4
    ),
    opacity = 0.66,
    name='Food & Entertaining',
    text = foodentertaining['HH_ID']
)

trace5 = Scatter(
    x=X, y=healthbeauty['Dollars'],
    line=Line(
        color='B4DDAF)',
        width=4
    ),
    opacity = 0.66,
    name='Health & Beauty',
    text = healthbeauty['HH_ID']
)

trace6 = Scatter(
    x=X, y=homegarden['Dollars'],
    line=Line(
        color='FF7FA5',
        width=4
    ),
    opacity = 0.66,
    name='Home & Garden',
    text = homegarden['HH_ID']
)

trace7 = Scatter(
    x=X, y=kidsbaby['Dollars'],
    line=Line(
        color='7FAEE2',
        width=4
    ),
    opacity = 0.66,
    name='Kids & Baby',
    text = kidsbaby['HH_ID']
)

trace8 = Scatter(
    x=X, y=teens['Dollars'],
    line=Line(
        color='E79A06',
        width=4
    ),
    opacity = 0.66,
    name='Teens',
    text = teens['HH_ID']
)

trace9 = Scatter(
    x=X, y=travelsports['Dollars'],
    line=Line(
        color='9276E3',
        width=4
    ),
    opacity = 0.66,
    name='Travel & Sports',
    text = travelsports['HH_ID']
)

trace10 = Scatter(
    x=X, y=others['Dollars'],
    line=Line(
        color='B0ADB0',
        width=4
    ),
    opacity = 0.66,
    name='Miscellaneous',
    text = others['HH_ID']
)


data = Data([trace1, trace2, trace3, trace4, trace5, trace6, trace7, trace8, trace9, trace10])
layout = Layout(
    title='Top Household Spending by Area',
    updatemenus=list([
        dict(
            x=-0.05,
            y=1,
            yanchor='top',
            buttons=list([
                dict(
                    args=['visible', [True, True, True, True, True, True, True, True, True, True]],
                    label='All',
                    method='restyle'
                ),
                dict(
                    args=['visible', [True, False, False, False, False, False, False, False, False, False]],
                    label='Electronics',
                    method='restyle'
                ),
                dict(
                    args=['visible', [False, True, False, False, False, False, False, False, False, False]],
                    label='Entertainment',
                    method='restyle'
                ),
                dict(
                    args=['visible', [False, False, True, False, False, False, False, False, False, False]],
                    label='Fashion',
                    method='restyle'
                ),
                dict(
                    args=['visible', [False, False, False, True, False, False, False, False, False, False]],
                    label='Food & Entertaining',
                    method='restyle'
                ),
                dict(
                    args=['visible', [False, False, False, False, True, False, False, False, False, False]],
                    label='Health & Beauty',
                    method='restyle'
                ),
                dict(
                    args=['visible', [False, False, False, False, False, True, False, False, False, False]],
                    label='Home & Garden',
                    method='restyle'
                ),
                dict(
                    args=['visible', [False, False, False, False, False, False, True, False, False, False]],
                    label='Kids & Baby',
                    method='restyle'
                ),
                dict(
                    args=['visible', [False, False, False, False, False, False, False, True, False, False]],
                    label='Teens',
                    method='restyle'
                ),
                dict(
                    args=['visible', [False, False, False, False, False, False, False, False, True, False]],
                    label='Travel & Sports',
                    method='restyle'
                ),
                dict(
                    args=['visible', [False, False, False, False, False, False, False, False, False, True]],
                    label='Miscellaneous',
                    method='restyle'
                )
            ]),
        )
    ]),
)
fig = Figure(data=data, layout=layout)
py.iplot(fig)

#Plot Revenue Sources by Product Area for company36
values = area.loc[area['CompanyID']==36].as_matrix()[0][1:]
labels = list(area)[1:]
fig = {"data": [{
        "values": values,
        "labels": labels,
    
        "domain": {"x": [0,5]},
        "name": "Product Areas",
        "hoverinfo": "label+percent+name",
        "hole": .4,
        "type": "pie"
}    
],
      "layout": {
          "title": "Company Products Composition",
          "annotations": [
          {
                "font": {
                    "size": 20
                },
                "showarrow": False,
                "text": "AREA",
                "x": 0.5,
                "y": 0.5
            }]
      }}
py.iplot(fig)

#Plot Revenue Sources by Major Categories for company36
values = category.loc[category['CompanyID']==36].as_matrix()[0][1:]
labels = list(category)[1:]
fig = {"data": [{
        "values": values,
        "labels": labels,
    
        "domain": {"x": [0,5]},
        "name": "Major Categories",
        "hoverinfo": "label+percent+name",
        "hole": .4,
        "type": "pie"
}    
],
      "layout": {
          "title": "Company Products Composition",
          "annotations": [
          {
                "font": {
                    "size": 20
                },
                "showarrow": False,
                "text": "CATEGORY",
                "x": 0.5,
                "y": 0.5
            }]
      }}
py.iplot(fig)

**Recommender Based On Collaborative Filtering Algorithms**

#Load Result Files
summary_actual_with_mean_cf_basics_cocluster_dol_log= pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A2_Data_CF_new/summary_actual_with_mean_cf_basics_cocluster_dol_log.csv")
summary_actual_with_mean_cf_basics_knn_dol_log= pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A2_Data_CF_new/summary_actual_with_mean_cf_basics_knn_dol_log.csv")
summary_actual_with_mean_cf_basics_svd_dol_log= pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A2_Data_CF_new/summary_actual_with_mean_cf_basics_svd_dol_log.csv")

summary_actual_with_mean_cf_basics_cocluster_dol_logscale= pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A2_Data_CF_new/summary_actual_with_mean_cf_basics_cocluster_dol_logscale_.csv")
summary_actual_with_mean_cf_basics_knn_dol_logscale= pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A2_Data_CF_new/summary_actual_with_mean_cf_basics_knn_dol_logscale_.csv")
summary_actual_with_mean_cf_basics_svd_dol_logscale= pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A2_Data_CF_new/summary_actual_with_mean_cf_basics_svd_dol_logscale_.csv")

summary_with_mean_cf_basics_knn_dol= pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A2_Data_CF_new/summary_with_mean_cf_basics_knn_dol.csv")
summary_with_mean_cf_basics_svd_dol= pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A2_Data_CF_new/summary_with_mean_cf_basics_svd_dol.csv")
summary_with_mean_cf_basics_cocluster_dol= pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A2_Data_CF_new/summary_with_mean_cf_basics_cocluster_dol.csv")

summary_actual_with_mean_cf_basics_cocluster_dol_scale= pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A2_Data_CF_new/summary_actual_with_mean_cf_basics_cocluster_dolscale_.csv")
summary_actual_with_mean_cf_basics_knn_dol_scale= pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A2_Data_CF_new/summary_actual_with_mean_cf_basics_knn_dolscale_.csv")
summary_actual_with_mean_cf_basics_svd_dol_scale= pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A2_Data_CF_new/summary_actual_with_mean_cf_basics_svd_dolscale_.csv")

#Plot Actual Mean Revenue trend of different algorithms for each decile ranked by _dol
trace1 = go.Scatter(
    x = summary_with_mean_cf_basics_knn_dol['percentile'],
    y = summary_with_mean_cf_basics_knn_dol['ActualRevenueMean'],
    name = 'knn_dol'
)
trace2 = go.Scatter(
    x = summary_with_mean_cf_basics_svd_dol['percentile'],
    y = summary_with_mean_cf_basics_svd_dol['ActualRevenueMean'],
    name = 'svd_dol'
)

trace3 = go.Scatter(
    x = summary_with_mean_cf_basics_cocluster_dol['percentile'],
    y = summary_with_mean_cf_basics_cocluster_dol['ActualRevenueMean'],
    name = 'cocluster_dol'
)

data = [trace1, trace2, trace3]

layout = dict(
              xaxis = dict(title = 'Top n percentile of predicted households ranked by score predicted'),
              yaxis = dict(title = 'Actual Mean Revenue'),
              )

fig = dict(data=data, layout=layout)
py.iplot(fig, filename='revenue_diff_measures')

#Plot Actual Mean Revenue trend of different algorithms for each decile ranked by _dol_log
trace4 = go.Scatter(
    x = summary_actual_with_mean_cf_basics_knn_dol_log['percentile'],
    y = summary_actual_with_mean_cf_basics_knn_dol_log['ActualRevenueMean'],
    name = 'knn_dol_log'
)
trace5 = go.Scatter(
    x = summary_actual_with_mean_cf_basics_svd_dol_log['percentile'],
    y = summary_actual_with_mean_cf_basics_svd_dol_log['ActualRevenueMean'],
    name = 'svd_dol_log'
)
trace6 = go.Scatter(
    x = summary_actual_with_mean_cf_basics_cocluster_dol_log['percentile'],
    y = summary_actual_with_mean_cf_basics_cocluster_dol_log['ActualRevenueMean'],
    name = 'cocluster_dol_log'
)


data = [trace4, trace5, trace6]

layout = dict(
              xaxis = dict(title = 'Top n percentile of predicted households ranked by score predicted'),
              yaxis = dict(title = 'Actual Mean Revenue'),
              )

fig = dict(data=data, layout=layout)
py.iplot(fig, filename='log_revenue_diff_measures')

#Plot Actual Mean Revenue trend of different algorithms for each decile ranked by _dol_scale
trace7 = go.Scatter(
    x = summary_actual_with_mean_cf_basics_knn_dol_scale['percentile'],
    y = summary_actual_with_mean_cf_basics_knn_dol_scale['ActualRevenueMean'],
    name = 'knn_dol_scale'
)
trace8 = go.Scatter(
    x = summary_actual_with_mean_cf_basics_svd_dol_scale['percentile'],
    y = summary_actual_with_mean_cf_basics_svd_dol_scale['ActualRevenueMean'],
    name = 'svd_dol_scale'
)

trace9 = go.Scatter(
    x = summary_actual_with_mean_cf_basics_cocluster_dol_scale['percentile'],
    y = summary_actual_with_mean_cf_basics_cocluster_dol_scale['ActualRevenueMean'],
    name = 'cocluster_dol_scale'
)

data = [trace7, trace8, trace9]

layout = dict(
              xaxis = dict(title = 'Top n percentile of predicted households ranked by score predicted'),
              yaxis = dict(title = 'Actual Mean Revenue'),
              )

fig = dict(data=data, layout=layout)
py.iplot(fig, filename='scale_revenue_diff_measures')

#Plot Atual Mean Revenue trend of different algorithms for each decile ranked by _dol_log_scale
trace10 = go.Scatter(
    x = summary_actual_with_mean_cf_basics_knn_dol_logscale['percentile'],
    y = summary_actual_with_mean_cf_basics_knn_dol_logscale['ActualRevenueMean'],
    name = 'knn_dol_log_scale'
)
trace11 = go.Scatter(
    x = summary_actual_with_mean_cf_basics_svd_dol_logscale['percentile'],
    y = summary_actual_with_mean_cf_basics_svd_dol_logscale['ActualRevenueMean'],
    name = 'svd_dol_log_scale'
)
trace12 = go.Scatter(
    x = summary_actual_with_mean_cf_basics_cocluster_dol_logscale['percentile'],
    y = summary_actual_with_mean_cf_basics_cocluster_dol_logscale['ActualRevenueMean'],
    name = 'cocluster_dol_log_scale'
)


data = [trace10, trace11, trace12]

layout = dict(
              xaxis = dict(title = 'Top n percentile of predicted households ranked by score predicted'),
              yaxis = dict(title = 'Actual Mean Revenue'),
              )

fig = dict(data=data, layout=layout)
py.iplot(fig, filename='log_scale_revenue_diff_measures')

#Plot Actual Mean Revenue trend of knn for each decile ranked by different score types
data_knn = [trace1, trace4, trace7, trace10]

layout = dict(
              xaxis = dict(title = 'Top n percentile of predicted households ranked by score predicted'),
              yaxis = dict(title = 'Actual Mean Revenue'),
              )

fig = dict(data=data_knn, layout=layout)
py.iplot(fig, filename='knn')

#Plot Actual Mean Revenue trend of svd for each decile ranked by different score types
data_svd = [trace2, trace5, trace8, trace11]

layout = dict(
              xaxis = dict(title = 'Top n percentile of predicted households ranked by score predicted'),
              yaxis = dict(title = 'Actual Mean Revenue'),
              )

fig = dict(data=data_svd, layout=layout)
py.iplot(fig, filename='svd')

#Plot Actual Mean Revenue trend of coclustering for each decile ranked by different score types
data_cocluster = [trace3, trace6, trace9, trace12]

layout = dict(
              xaxis = dict(title = 'Top n percentile of predicted households ranked by score predicted'),
              yaxis = dict(title = 'Actual Mean Revenue'),
              )

fig = dict(data=data_cocluster, layout=layout)
py.iplot(fig, filename='cocluster')

## SVD

#Plot Actual Mean Revenue trend of SVD for each decile ranked by different score types of revenue and number
summary_cf_further_svd_dol = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A5_lastmin/summary_cf_further_svd_dol.csv")
summary_cf_further_svd_dol_log = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A5_lastmin/summary_cf_further_svd_dol_log.csv")
summary_cf_further_svd_num = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A5_lastmin/summary_cf_further_svd_num.csv")
summary_cf_further_svd_num_log = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A5_lastmin/summary_cf_further_svd_num_log.csv")

tracef1 = go.Scatter(
    x = summary_cf_further_svd_dol['percentile'],
    y = summary_cf_further_svd_dol['ActualRevenueMean'],
    name = 'svd_dol'
)
tracef2 = go.Scatter(
    x = summary_cf_further_svd_dol_log['percentile'],
    y = summary_cf_further_svd_dol_log['ActualRevenueMean'],
    name = 'svd_dol_log'
)
tracef3 = go.Scatter(
    x = summary_cf_further_svd_num['percentile'],
    y = summary_cf_further_svd_num['ActualRevenueMean'],
    name = 'svd_num'
)
tracef4 = go.Scatter(
    x = summary_cf_further_svd_num_log['percentile'],
    y = summary_cf_further_svd_num_log['ActualRevenueMean'],
    name = 'svd_num_log'
)

data = [tracef1, tracef2, tracef3, tracef4]

layout = dict(
              xaxis = dict(title = 'Top n percentile of predicted households ranked by score predicted'),
              yaxis = dict(title = 'Actual Mean Revenue'),
              )

fig = dict(data=data, layout=layout)
py.iplot(fig, filename='svd_further')

## KNN

#Load result files for KNN
import os
path = 'F:/AY16-17 Sem2/BT4221/Project/A6_13lines'
knnFiles = [f for f in os.listdir(path) if("summary_knn_company_36_" in f)]

#Construct trace data
list_ = []
for file in knnFiles:
    list_.append(pd.read_csv(path+"/"+file))

data = []
i = 0
for file in list_:
    trace = go.Scatter(x = file['percentile'],y = file['ActualRevenueMean'],name = knnFiles[i].replace("summary_knn_company_36_",'').replace(".csv",''))
    data.append(trace)
    i+=1

#Construct the dropdown layout
plot_list = []
def boo_label (label):
    if (label =='dol' or label=='num'):
        list_bool = [((label in name)&("log"not in name)) for name in knnFiles]
    elif label =='pearson':
        list_bool = [((label in name)&("baseline"not in name)) for name in knnFiles]
    else:
        list_bool = [(label in name) for name in knnFiles]
    return (list_bool)

labels = ["dol", "num", "dol_log", "num_log", "msd","cosine","pearson","pearson_baseline"]
for label in labels:
    plot_list.append(dict(method='restyle',label=label,args=['visible',boo_label (label) ]))
        
buttons =   list([
                dict(
                    args=['visible', [True, True, True, True, True, True, True, True, True, True, True, True, True]],
                    label='All',
                    method='restyle'
                )])+   plot_list   

#Plot Actual Mean Revenue trend of KNN using different similarity measures for each decile ranked by different score types
layout = go.Layout(
    title='Comparing KNN by Score and Similarity',
    updatemenus=list([
        dict(
            x=-0.05,
            y=1,
            yanchor='top',
            buttons= buttons)
    ]),
)
fig = go.Figure(data=data, layout=layout)
py.iplot(fig)

**Popularity-Based Recommender**

#Plot Actual Mean Revenue trend of popularity, popularity_by_area, and popualrity_by_category
popularity_all = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A1_Code_Data_Descriptive_PopularityRec/summary_popularity_company_36.csv")
popularity_area = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A1_Code_Data_Descriptive_PopularityRec/summary_popularity_company_36_area.csv")
popularity_cat = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A1_Code_Data_Descriptive_PopularityRec/summary_popularity_company_36_category.csv")

tracep1 = go.Scatter(
    x = popularity_all['percentile'],
    y = popularity_all['ActualRevenue'],
    name = 'Popularity_all'
)
tracep2 = go.Scatter(
    x = popularity_area['percentile'],
    y = popularity_area['ActualRevenue'],
    name = 'Popularity_area'
)
tracep3 = go.Scatter(
    x = popularity_cat['percentile'],
    y = popularity_cat['ActualRevenue'],
    name = 'Popularity_category'
)

datap = [tracep1, tracep2, tracep3]

layout = dict(
              xaxis = dict(title = 'Top n percentile of predicted households ranked by household spending'),
              yaxis = dict(title = 'Actual Mean Revenue'),
              )

fig = dict(data=datap, layout=layout)
py.iplot(fig, filename='revenue_popularity')

#Compare the performance of 3 best strategies
num_sim_cosine=pd.read_csv(path+"/"+knnFiles[10])
tracek = go.Scatter(
    x = num_sim_cosine['percentile'],
    y = num_sim_cosine['ActualRevenueMean'],
    name = 'knn_num_sim_cosine'
)
data_k = [trace5, tracek, tracep2]
layout = dict(
              xaxis = dict(title = 'Top n percentile of predicted households ranked by household spending'),
              yaxis = dict(title = 'Actual Mean Revenue'),
              )

fig = dict(data=data_k, layout=layout)
py.iplot(fig, filename='comparing')

X1 = list(range(100))
X2 = list(range(200))
X3 = list(range(500))
X4 = list(range(1000))

#Load Recommendation File of Popularity-based recommender
alld = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A1_Code_Data_Descriptive_PopularityRec/popularity_company_36.csv")
area = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A1_Code_Data_Descriptive_PopularityRec/popularity_company_36_area.csv")
cat = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/A1_Code_Data_Descriptive_PopularityRec/popularity_company_36_category.csv")

d_all = alld['ActualRevenue']
d_area = area['ActualRevenue']
d_cat = cat['ActualRevenue']

h_all = alld['HH_ID']
h_area = area['HH_ID']
h_cat = cat['HH_ID']

alld1 = alld[:1000]
area1 = area[:1000]
cat1 = cat[:1000]

alld1['cumulative_revenue'] = [sum(alld1['ActualRevenue'][:i]) for i in range(1, len(alld1['ActualRevenue'])+1)]
area1['cumulative_revenue'] = [sum(area1['ActualRevenue'][:i]) for i in range(1, len(area1['ActualRevenue'])+1)]
cat1['cumulative_revenue'] = [sum(cat1['ActualRevenue'][:i]) for i in range(1, len(cat1['ActualRevenue'])+1)]

c_all = alld1['cumulative_revenue']
c_area = area1['cumulative_revenue']
c_cat = cat1['cumulative_revenue']

#Plot actual household spending of top 100/1000 customers recommended by popularity
trace1 = go.Bar(
    x = X1,
    y = d_all[:99],
    name = '100',
    text = h_all[:99]
)

trace2 = go.Bar(
    x = X4,
    y = d_all[:999],
    name = '1000',
    text = h_all[:999]
)

trace11 = go.Scatter(
    x = X1,
    y = c_all[:99],
    mode = 'lines',
    name = '100',
    yaxis='y2'
)

trace21 = go.Scatter(
    x = X4,
    y = c_all[:999],
    mode = 'lines',
    name = '1000',
    yaxis='y2'
)

data = [trace1, trace2, trace11, trace21]
layout = go.Layout(
    title = 'Popularity',
    yaxis=dict(
        title='Household Spending'
    ),
    yaxis2=dict(
        title='Cumulative Spending',
        titlefont=dict(
            color='rgb(148, 103, 189)'
        ),
        tickfont=dict(
            color='rgb(148, 103, 189)'
        ),
        overlaying='y',
        side='right'),
    updatemenus=list([
        dict(
            x=-0.05,
            y=1,
            yanchor='top',
            buttons=list([
                dict(
                    args=['visible', [True, False, True, False]],
                    label='top 100',
                    method='restyle'
                ),
                dict(
                    args=['visible', [False, True,False, True]],
                    label='top 1000',
                    method='restyle'
                )
    ]),)]),
)

fig = go.Figure(data=data, layout=layout)
py.iplot(fig)

#Plot actual household spending of top 100/1000 customers recommended by customerised popularity-area
trace1 = go.Bar(
    x = X1,
    y = d_area[:99],
    name = '100',
    text = h_area[:99],
    marker=dict(
        color='rgb(204,204,204)',
    )
)

trace2 = go.Bar(
    x = X4,
    y = d_area[:999],
    name = '1000',
    text = h_area[:999],
    marker=dict(
        color='rgb(204,204,204)',
    )
)

trace11 = go.Scatter(
    x = X1,
    y = c_area[:99],
    mode = 'lines',
    name = '100',
    yaxis='y2'
)

trace21 = go.Scatter(
    x = X4,
    y = c_area[:999],
    mode = 'lines',
    name = '1000',
    yaxis='y2'
)

data = [trace1, trace2, trace11, trace21]
layout = go.Layout(
    title = 'Customerised Popularity - By Area',
    yaxis=dict(
        title='Household Spending'
    ),
    yaxis2=dict(
        title='Cumulative Spending',
        titlefont=dict(
            color='rgb(148, 103, 189)'
        ),
        tickfont=dict(
            color='rgb(148, 103, 189)'
        ),
        overlaying='y',
        side='right'),
    updatemenus=list([
        dict(
            x=-0.05,
            y=1,
            yanchor='top',
            buttons=list([
                dict(
                    args=['visible', [True, False, True, False]],
                    label='top 100',
                    method='restyle'
                ),
                dict(
                    args=['visible', [False, True,False, True]],
                    label='top 1000',
                    method='restyle'
                )
    ]),)]),
)

fig = go.Figure(data=data, layout=layout)
py.iplot(fig)

#Plot actual household spending of top 100/1000 customers recommended by customerised popularity-category
trace1 = go.Bar(
    x = X1,
    y = d_cat[:99],
    name = '100',
    text = h_cat[:99],
    marker=dict(
        color='rgb(158,202,225)'),
    opacity=0.7
)

trace2 = go.Bar(
    x = X4,
    y = d_cat[:999],
    name = '1000',
    text = h_cat[:999],
    marker=dict(
        color='rgb(158,202,225)'),
    opacity=0.7
)

trace11 = go.Scatter(
    x = X1,
    y = c_cat[:99],
    mode = 'lines',
    name = '100',
    yaxis='y2'
)

trace21 = go.Scatter(
    x = X4,
    y = c_cat[:999],
    mode = 'lines',
    name = '1000',
    yaxis='y2'
)

data = [trace1, trace2, trace11, trace21]
layout = go.Layout(
    title = 'Customerised Popularity - By Category',
    yaxis=dict(
        title='Household Spending'
    ),
    yaxis2=dict(
        title='Cumulative Spending',
        titlefont=dict(
            color='rgb(148, 103, 189)'
        ),
        tickfont=dict(
            color='rgb(148, 103, 189)'
        ),
        overlaying='y',
        side='right'),
    updatemenus=list([
        dict(
            x=-0.05,
            y=1,
            yanchor='top',
            buttons=list([
                dict(
                    args=['visible', [True, False, True, False]],
                    label='top 100',
                    method='restyle'
                ),
                dict(
                    args=['visible', [False, True,False, True]],
                    label='top 1000',
                    method='restyle'
                )
    ]),)]),
)

fig = go.Figure(data=data, layout=layout)
py.iplot(fig)

#Heatmap of potential customer distribution in the US by zipcode, data from customerised popularity recommender result-by category
popzip = pd.read_csv("F:/AY16-17 Sem2/BT4221/Project/hhhhhh/3.popularity of hh by spending/pop_zip_cat.csv")

scl = [[0.0, 'rgb(242,240,247)'],[0.2, 'rgb(218,218,235)'],[0.4, 'rgb(188,189,220)'],\
            [0.6, 'rgb(158,154,200)'],[0.8, 'rgb(117,107,177)'],[1.0, 'rgb(84,39,143)']]
data = [ dict(
        type='choropleth',
        colorscale = scl,
        autocolorscale = False,
        locations = popzip['State'],
        z = popzip['Dollars'].astype(float),
        locationmode = 'USA-states',
        marker = dict(
            line = dict (
                color = 'rgb(255,255,255)',
                width = 2
            ) ),
        colorbar = dict(
            title = "USD")
        ) ]

layout = dict(
        title = 'Customised Popularity - By Category',
        geo = dict(
            scope='usa',
            projection=dict( type='albers usa' ),
            showlakes = True,
            lakecolor = 'rgb(255, 255, 255)'),
             )
    
fig = dict( data=data, layout=layout )
py.iplot( fig, filename='coloropleth-map' )
